backend/requirements.txt
fastapi==0.128.7
uvicorn==0.40.0
playwright==1.58.0
feedparser==6.0.12
httpx==0.28.1
requests==2.32.5
beautifulsoup4==4.14.3
pydantic==2.12.5
python-dotenv==1.2.1
tenacity==9.1.4
sqlalchemy==2.0.44
psycopg[binary]==3.2.12
alembic==1.16.5
